remove opss only journey
