+++ config/routes.rb
@@ -75,9 +75,9 @@ Rails.application.routes.draw do
 
   resources :create_a_case_page, controller: "create_a_case_page", only: %i[index]
 
-  resources :enquiry, controller: "investigations/enquiry", only: %i[show new create update]
-  resources :allegation, controller: "investigations/allegation", only: %i[show new create update]
-  resources :project, controller: "investigations/project", only: %i[show new create update]
+  resources :enquiry, controller: "investigations/enquiry", only: %i[show update]
+  resources :allegation, controller: "investigations/allegation", only: %i[show update]
+  resources :project, controller: "investigations/project", only: %i[show update]
   resources :ts_investigation, controller: "investigations/ts_investigations", only: %i[show new create update]
 
   scope :investigation, path: "", as: :investigation do
@@ -88,10 +88,10 @@ Rails.application.routes.draw do
   end
 
   scope :investigation, path: "", module: "investigations", as: :investigation do
-    resources :enquiry,          controller: "enquiry",           only: %i[show new create update]
-    resources :allegation,       controller: "allegation",        only: %i[show new create update]
-    resources :project,          controller: "project",           only: %i[show new create update]
-    resources :ts_investigation, controller: "ts_investigations", only: %i[show new create update]
+    resources :enquiry,          controller: "enquiry",           only: %i[show update]
+    resources :allegation,       controller: "allegation",        only: %i[show update]
+    resources :project,          controller: "project",           only: %i[show update]
+    resources :ts_investigation, controller: "ts_investigations", only: %i[show update]
   end
 
   resource :investigations, only: [], path: "cases" do
@@ -103,9 +103,11 @@ Rails.application.routes.draw do
     get "all-cases", to: "investigations#index"
   end
 
+  get "cases/new", to: "ts_investigations#new"
+
   resources :investigations,
             path: "cases",
-            only: %i[show new index destroy],
+            only: %i[show index destroy],
             param: :pretty_id,
             concerns: %i[document_attachable] do
     member do
